- Fix a problem with utils.LoadData & utils.ConvertData causing go data structure returned by HCL to not be flatten - Always use default delimiters for gotemplate extensions
